moved authentication logic out of api.py
